+++ src-ui/js/ui/ToolArea.js
@@ -352,6 +352,9 @@ ui.toolarea = {
 	flushexcell: function() {
 		ui.puzzle.board.flushexcell();
 	},
+	traceroute: function() {
+		ui.puzzle.board.operate("traceroute");
+	},
 	applypreset: function(e) {
 		ui.auxeditor.close();